twitterJob/requirements.txt
tweepy
pandas
psycopg2-binary
sqlalchemy
openpyxl

